Move convert form into separate component and initialize session id before rendering page
