+++ apps/cinewrax/src/app/convert/page.tsx
@@ -1,96 +1,8 @@
-"use client";
-
-import {
-  Box,
-  Button,
-  FormControl,
-  FormControlLabel,
-  Radio,
-  RadioGroup,
-  Slider,
-  Switch,
-  TextField,
-  Typography,
-} from "@mui/material";
-import { ChangeEventHandler, useState } from "react";
-
-import { useSessionId } from "../../hooks/useSessionId";
+import { getSessionId } from "../../utils/session";
+import { ConvertForm } from "./components/ConvertForm/ConvertForm";
 
 export default function Convert() {
-  const sessionId = useSessionId();
+  const sessionId = getSessionId();
 
-  const [audioFormat, setAudioFormat] = useState("MP3");
-  const [bitrate, setBitrate] = useState(128);
-  const [file, setFile] = useState<null | File>(null);
-
-  const handleAudioFormatChange: ChangeEventHandler<HTMLInputElement> = (event) => {
-    setAudioFormat(event.target.value);
-  };
-
-  const handleBitrateChange = (event: Event, newValue: number | number[]) => {
-    setBitrate(newValue as number);
-  };
-
-  const handleFileChange: ChangeEventHandler<HTMLInputElement> = (event) => {
-    if (event.target.files === null || event.target.files.length === 0) {
-      setFile(null);
-      return;
-    }
-
-    setFile(event.target.files[0]);
-  };
-
-  return (
-    <Box sx={{ maxWidth: 400, margin: "auto", padding: 4 }}>
-      <Typography variant="h4" component="h1" gutterBottom>
-        Audio Converter
-      </Typography>
-
-      {/* File Input */}
-      <FormControl fullWidth sx={{ mb: 2 }}>
-        <TextField
-          type="file"
-          onChange={handleFileChange}
-          inputProps={{ accept: "audio/*" }}
-          helperText={file ? `Selected: ${file.name}` : "Choose an audio file"}
-        />
-      </FormControl>
-
-      {/* Toggle for Audio Format */}
-      <FormControl component="fieldset">
-        <Typography component="legend">Select Audio Format</Typography>
-        <RadioGroup
-          aria-label="audio-format"
-          name="audio-format"
-          value={audioFormat}
-          onChange={handleAudioFormatChange}
-          row
-        >
-          <FormControlLabel value="MP3" control={<Radio />} label="MP3" />
-          <FormControlLabel value="WAV" control={<Radio />} label="WAV" />
-        </RadioGroup>
-      </FormControl>
-
-      {/* Bitrate Slider for MP3 */}
-      {audioFormat === "MP3" && (
-        <Box sx={{ mt: 4 }}>
-          <Typography gutterBottom>Bitrate: {bitrate} kbps</Typography>
-          <Slider value={bitrate} min={64} max={320} step={1} onChange={handleBitrateChange} valueLabelDisplay="auto" />
-        </Box>
-      )}
-
-      {/* Submit Button */}
-      <Button
-        variant="contained"
-        color="primary"
-        fullWidth
-        sx={{ mt: 4 }}
-        onClick={() => {
-          console.log(`Audio format: ${audioFormat}, Bitrate: ${bitrate} kbps`);
-        }}
-      >
-        Convert Audio
-      </Button>
-    </Box>
-  );
+  return <ConvertForm sessionId={sessionId} />;
 }
